requirements.txt
py-moneyed==2.0
pandas==1.4.0
fastapi==0.72.0
uvicorn==0.15.0